Implemented polling the controller for state. I might change this implementation around a bit...
Right now the Java side is close to the native steamapi library:
you pass in a pre-allocated structure and it's populated with the controller data.
I keep a struct around on the native side, populate that, and then call a
Java method to populate the Java object.

Maybe there's an easier way. Maybe not.

Oh... I bet there are some threading issues I should be aware of...