Ok, so querying a lot of these controller methods when not running inside Steam causes a segfault.
We don't like segfaults happening in our nice Java code.

Protect all of these methods from actually calling into the native side
unless properly initialized, saving users from having to wrap every call
inside some sort of are-we-on-steam check.
Now, maybe the are-we-on-steam check isn't onerous and perhaps in the
end a game's interaction with this will happen in a class that is
only even instantiated if we're on steam. If so I can roll this back.

Also:
- Added a hasController(int index) for checking to see if a controller exists
  without populating a State object.
- Switch to "nMethodName" instead of "nativeMethodName" native naming scheme.