Let's "ship" a 1.2 release.
Instead of 1.1.x just to show that we're not fooling around. I'm sure many
1.2.x releases will follow as I iron out kinks later.

The reason for air quotes around ship is that shipping now means setting the
version, installing the Maven artifacts into this repository, and committing
them. Stone tools! But unfortunately necessary because we need to commit the
Steamworks SDK libraries that go along with our native libraries and can't rely
on the depending project git-submoduling us into position and doing a local
install during their build as the Steamworks SDK is not available at that
point.

But we're trying to avoid having a magic secret Maven repository behind a VPN,
which makes life extra complicated. Instead we have a bunch of private Github
projects and our CI build (running on Github) has credentials that allow it to
access all those projects. One stop shopping!